meu-projeto-edu/backend/requirements.txt
Django==5.2.14
django-ninja==1.6.2
django-ninja-jwt==5.4.4
pydantic==2.13.4
psycopg2-binary==2.9.9
dj-database-url==2.2.0